refactor: extract saved search packages
