+++ lib/features/immersion/list/immersion_list_actions.dart
@@ -1,5 +1,4 @@
 import 'package:pass_emploi_app/models/immersion.dart';
-import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
 import 'package:pass_emploi_app/models/location.dart';
 
 class ImmersionListRequestAction {
@@ -22,14 +21,3 @@ class ImmersionListFailureAction {}
 
 class ImmersionListResetAction {}
 
-class SavedImmersionSearchRequestAction {
-  final String codeRome;
-  final Location location;
-  final ImmersionSearchParametersFiltres filtres;
-
-  SavedImmersionSearchRequestAction({
-    required this.codeRome,
-    required this.location,
-    required this.filtres,
-  });
-}
