+++ lib/features/immersion/parameters/immersion_search_parameters_reducer.dart
@@ -1,6 +1,7 @@
 import 'package:pass_emploi_app/features/immersion/list/immersion_list_actions.dart';
 import 'package:pass_emploi_app/features/immersion/parameters/immersion_search_parameters_actions.dart';
 import 'package:pass_emploi_app/features/immersion/parameters/immersion_search_parameters_state.dart';
+import 'package:pass_emploi_app/features/immersion/saved_search/immersion_saved_search_actions.dart';
 import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
 
 ImmersionSearchParametersState immersionSearchParametersState(
