+++ lib/features/offre_emploi/saved_search/offre_emploi_saved_search_actions.dart
@@ -1,0 +1,16 @@
+import 'package:pass_emploi_app/models/location.dart';
+import 'package:pass_emploi_app/models/offre_emploi_filtres_parameters.dart';
+
+class SavedOffreEmploiSearchRequestAction {
+  final String keywords;
+  final Location? location;
+  final bool onlyAlternance;
+  final OffreEmploiSearchParametersFiltres filtres;
+
+  SavedOffreEmploiSearchRequestAction({
+    required this.keywords,
+    required this.location,
+    required this.onlyAlternance,
+    required this.filtres,
+  });
+}
