+++ lib/features/offre_emploi/search/offre_emploi_search_actions.dart
@@ -1,6 +1,5 @@
 import 'package:pass_emploi_app/models/location.dart';
 import 'package:pass_emploi_app/models/offre_emploi.dart';
-import 'package:pass_emploi_app/models/offre_emploi_filtres_parameters.dart';
 
 class OffreEmploiSearchRequestAction {
   final String keywords;
@@ -25,17 +24,3 @@ class OffreEmploiSearchFailureAction {}
 class OffreEmploiSearchRequestMoreResultsAction {}
 
 class OffreEmploiSearchResetAction {}
-
-class SavedOffreEmploiSearchRequestAction {
-  final String keywords;
-  final Location? location;
-  final bool onlyAlternance;
-  final OffreEmploiSearchParametersFiltres filtres;
-
-  SavedOffreEmploiSearchRequestAction({
-    required this.keywords,
-    required this.location,
-    required this.onlyAlternance,
-    required this.filtres,
-  });
-}
