+++ lib/features/offre_emploi/search/offre_emploi_search_middleware.dart
@@ -63,18 +63,6 @@ class OffreEmploiMiddleware extends MiddlewareClass<AppState> {
             filtres: action.updatedFiltres,
           ),
         );
-      } else if (action is SavedOffreEmploiSearchRequestAction) {
-        _resetSearchWithUpdatedFiltres(
-          store: store,
-          userId: userId,
-          request: SearchOffreEmploiRequest(
-            keywords: action.keywords,
-            location: action.location,
-            onlyAlternance: action.onlyAlternance,
-            page: 1,
-            filtres: action.filtres,
-          ),
-        );
       }
     }
   }
